Agrego FullCardContainer
    refs:#381
